requirements.txt
flask>=3.0.2
flask-cors>=4.0.0
langchain>=0.1.13
langchain-google-genai>=0.1.0
tavily-python>=0.3.4
python-dotenv>=1.0.0
requests>=2.31.0